Replace concrete query with deferred query
This is a tricky one. ActiveRecord#all evaluates immediately, returning
an Array. This _should_ be supported, but it is not actually how it
should be done.

By using #includes here, we are actually optimizing a couple
significant ways.
 1 - Lazy load, we will actually only load 1 page of sorted results
instead of the entire table
 2 - Intelligently pre-fetching the associated users associated with
the records (but only for the 1 page as per #1)

This is how the API _should_ be used, but not a very good test of
support for bad cases. We'll have to write an example using the
previous method to ensure we don't break it. Either that or detect/warn
poor usage like that.